+++ examples/ha-3nic/main.tf
@@ -12,7 +12,7 @@ terraform {
 # Create a firewall rule to allow BIG-IP ConfigSync
 module "ha_fw" {
   source                = "memes/f5-bigip/google//modules/configsync-fw"
-  version               = "1.3.1"
+  version               = "1.3.2"
   project_id            = var.project_id
   bigip_service_account = var.service_account
   dataplane_network     = var.internal_network
@@ -51,7 +51,7 @@ resource "google_compute_address" "int" {
 
 module "ha" {
   source                            = "memes/f5-bigip/google//modules/ha"
-  version                           = "1.3.1"
+  version                           = "1.3.2"
   project_id                        = var.project_id
   num_instances                     = var.num_instances
   zones                             = [var.zone]