+++ modules/cfe/main.tf
@@ -18,13 +18,16 @@ locals {
 
 # CFE deployment is simply HA with a custom script to apply CFE JSON
 module "ha" {
-  source                 = "./../ha/"
-  project_id             = var.project_id
-  zones                  = var.zones
-  num_instances          = var.num_instances
-  instance_name_template = var.instance_name_template
-  description            = var.description
-  metadata               = local.metadata
+  source                  = "./../ha/"
+  project_id              = var.project_id
+  zones                   = var.zones
+  num_instances           = var.num_instances
+  instance_name_template  = var.instance_name_template
+  instance_ordinal_offset = var.instance_ordinal_offset
+  domain_name             = var.domain_name
+  search_domains          = var.search_domains
+  description             = var.description
+  metadata                = local.metadata
   # Make sure the labels applied to instances have the CFE specific key-value pair
   labels                          = merge(var.labels, { "${var.cfe_label_key}" = var.cfe_label_value })
   tags                            = var.tags
@@ -64,6 +67,7 @@ module "ha" {
   default_gateway                   = var.default_gateway
   use_cloud_init                    = var.use_cloud_init
   admin_password_secret_manager_key = var.admin_password_secret_manager_key
+  secret_implementor                = var.secret_implementor
   as3_payloads                      = var.as3_payloads
   do_payloads                       = var.do_payloads
   install_cloud_libs                = var.install_cloud_libs